+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2019-05-07
+
+### Fixed
+- removed haste leftover
+- some namespaces
+
+### Changed
+- increased utils bundle dependency
+- small refactoring
+
 ## [1.0.3] - 2019-01-22
 
 ### Fixed